refactor: Move `isNotionErrorWithCode` into `notion-utils` directory
