+++ src/content/sync/notion-utils/index.ts
@@ -1,3 +1,4 @@
 export { buildDate } from './build-date';
 export { buildRichText } from './build-rich-text';
+export { isNotionErrorWithCode } from './error';
 export { convertWebURLToAppURL, getPageIDFromURL, isNotionURL } from './url';
